pyproject.toml
[project]
name = "dirassati"
version = "0.1.0"
description = ""
authors = [
    {name = "ttyxd",email = "cmax0890@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-restx (>=1.3.0,<2.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fakeredis (>=2.28.1,<3.0.0)",
    "flask-redis (>=0.4.0,<0.5.0)",
    "itsdangerous (>=2.1.2,<3.0.0)",
    "mailjet-rest (>=1.3.4,<2.0.0)",
    "flask-limiter (>=3.12,<4.0)",
    "chardet (>=5.2.0,<6.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
